chore: cleanup wallet connect
